Clean up main.rs.
Main.rs has been getting pretty unwieldy. It's now been extracted into a
few different modules, making the CLI path easier to read.
