+++ src/main.rs
@@ -12,6 +12,7 @@
 // this program. If not, see <https://www.gnu.org/licenses/>.
 //
 mod audio;
+mod cli;
 mod config;
 mod controller;
 mod dmx;
@@ -28,661 +29,17 @@ mod testutil;
 mod util;
 mod verify;
 
-use crate::playlist::Playlist;
-use clap::{crate_version, Parser, Subcommand};
-use lighting::parser::{parse_light_shows, utils::parse_time_string};
-use lighting::validation::validate_groups;
-use proto::player::v1::player_service_client::PlayerServiceClient;
-use proto::player::v1::{
-    GetActiveEffectsRequest, GetCuesRequest, NextRequest, PlayFromRequest, PlayRequest,
-    PreviousRequest, Song, StatusRequest, StopRequest, SwitchToPlaylistRequest,
-};
-use std::collections::HashSet;
-use std::env;
-use std::error::Error;
-use std::fmt::Display;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use std::time::Duration;
-use tonic::transport::Channel;
-use tonic::Request;
-use tracing::info;
-
-const SYSTEMD_SERVICE: &str = r#"
-[Unit]
-Description=multitrack player
-
-[Service]
-Type=simple
-Restart=on-failure
-EnvironmentFile=-/etc/default/mtrack
-ExecStart={{ CURRENT_EXECUTABLE }} start "$MTRACK_CONFIG"
-ExecReload=/bin/kill -HUP $MAINPID
-
-# User and group. Create with:
-#   sudo useradd --system --no-create-home --shell /usr/sbin/nologin mtrack
-#   sudo usermod -aG audio mtrack
-User=mtrack
-Group=mtrack
-SupplementaryGroups=audio
-
-# Allow setting thread/RT priority for real-time audio scheduling.
-AmbientCapabilities=CAP_SYS_NICE
-CapabilityBoundingSet=CAP_SYS_NICE
-
-# Filesystem restrictions. The entire filesystem is read-only, which is
-# sufficient since mtrack does not write to disk. /home is inaccessible.
-ProtectSystem=strict
-ProtectHome=true
-PrivateTmp=true
-
-# Kernel restrictions.
-ProtectKernelTunables=true
-ProtectKernelModules=true
-ProtectKernelLogs=true
-ProtectControlGroups=true
-
-# Additional hardening.
-NoNewPrivileges=true
-LockPersonality=true
-RestrictNamespaces=true
-RestrictSUIDSGID=true
-MemoryDenyWriteExecute=true
-SystemCallArchitectures=native
-RestrictAddressFamilies=AF_INET AF_INET6 AF_UNIX
-
-[Install]
-WantedBy=multi-user.target
-Alias=mtrack.service
-"#;
-
-/// Default thread priority on the [0; 100) scale when MTRACK_THREAD_PRIORITY is unset.
-/// On Linux (normal scheduling) this is roughly in the nice -8 to -10 range—higher than default 0 but below max.
-const DEFAULT_THREAD_PRIORITY: u8 = 70;
-
-/// Sets the current thread's priority for better audio scheduling (cross-platform).
-/// Uses the [0; 100) scale: higher value = higher priority.
-/// Default is 70 when MTRACK_THREAD_PRIORITY is unset; set to 0-99 to override.
-/// Fails silently if lacking permission. Use high values with care; they can starve other threads.
-fn apply_thread_priority() {
-    use thread_priority::{set_current_thread_priority, ThreadPriority, ThreadPriorityValue};
-
-    let priority = env::var("MTRACK_THREAD_PRIORITY")
-        .ok()
-        .and_then(|v| {
-            let n = v.parse::<u8>().ok()?;
-            (n < 100).then(|| ThreadPriorityValue::try_from(n).ok())?
-        })
-        .unwrap_or_else(|| ThreadPriorityValue::try_from(DEFAULT_THREAD_PRIORITY).unwrap());
-
-    match set_current_thread_priority(ThreadPriority::Crossplatform(priority)) {
-        Ok(()) => info!("Set thread priority for audio"),
-        Err(e) => tracing::warn!(
-            error = %e,
-            "Could not set thread priority (e.g. run as root or with CAP_SYS_NICE on Linux)"
-        ),
-    }
-}
-
-#[derive(Parser)]
-#[clap(
-    author = "Michael Wilson",
-    version = crate_version!(),
-    about = "A multitrack player."
-)]
-struct Cli {
-    #[clap(subcommand)]
-    command: Commands,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-    /// Lists and verifies all songs in the given directory.
-    Songs {
-        /// The path to the songs repository on disk.
-        path: String,
-        /// Initialize song directories with YAML files containing default values.
-        #[arg(long)]
-        init: bool,
-    },
-    /// Lists the available audio output devices.
-    Devices {},
-    /// Lists the available MIDI input/output devices.
-    MidiDevices {},
-    /// Plays a song through the audio interface.
-    /// Playlist will verify a playlist.
-    Playlist {
-        /// The path to the song repository.
-        repository_path: String,
-        /// The path to the playlist.
-        playlist_path: String,
-    },
-    /// Start will start the multitrack player.
-    Start {
-        /// The path to the player config.
-        player_path: String,
-        /// The path to the playlist. Must be specified if the playlist is not specified in the player config.
-        playlist_path: Option<String>,
-    },
-    /// Plays the current song in the playlist.
-    Play {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-        /// Start playback from a specific time (e.g., "1:23.456" or "45.5s").
-        #[arg(long)]
-        from: Option<String>,
-    },
-    /// Moves to the previous song in the playlist.
-    Previous {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Moves to the next song in the playlist.
-    Next {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Stops the currently playing song.
-    Stop {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Switches to the given playlist.
-    SwitchToPlaylist {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-        /// The name of the playlist to switch to. Currently only supports "all_songs" and "playlist."
-        playlist_name: String,
-    },
-    /// Gets the current status of the player from the gRPC server.
-    Status {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Prints all active lighting effects from the gRPC server.
-    ActiveEffects {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Prints a systemd service definition to stdout.
-    Systemd {},
-    /// Verifies the syntax of a light show file.
-    VerifyLightShow {
-        /// The path to the light show file to verify.
-        show_path: String,
-        /// Optional path to mtrack.yaml config file to validate group/fixture names.
-        #[arg(short, long)]
-        config: Option<String>,
-    },
-    /// Lists all cues in the current song's lighting timeline.
-    Cues {
-        /// The host and port of the gRPC server.
-        #[arg[short, long]]
-        host_port: Option<String>,
-    },
-    /// Verifies songs in a repository against the player config.
-    Verify {
-        /// The path to the mtrack.yaml player config file.
-        config: String,
-        /// Only check specific categories (e.g., "track-mappings"). Runs all checks if omitted.
-        #[arg(long)]
-        check: Option<Vec<String>>,
-        /// Hostname to verify against. When audio_profiles are used, this filters which profiles
-        /// to check. If omitted, all profiles are verified.
-        #[arg(long)]
-        hostname: Option<String>,
-    },
-}
-
-/// Prints a list of devices for the Devices and MidiDevices subcommands.
-fn print_device_list<T: Display>(devices: Vec<T>, empty_msg: &str) {
-    if devices.is_empty() {
-        println!("{}", empty_msg);
-        return;
-    }
-    println!("Devices:");
-    for device in devices {
-        println!("- {}", device);
-    }
-}
-
-/// Verifies a light show file, optionally validating against a config file.
-fn verify_light_show(show_path: &str, config_path: Option<&str>) -> Result<(), Box<dyn Error>> {
-    let path = Path::new(show_path);
-
-    if !path.exists() {
-        return Err(format!("Light show file not found: {}", show_path).into());
-    }
-
-    // Read and parse the light show
-    let content = std::fs::read_to_string(path)?;
-    let shows = match parse_light_shows(&content) {
-        Ok(shows) => shows,
-        Err(e) => {
-            eprintln!("❌ Syntax error in light show:");
-            eprintln!("{}", e);
-            return Err(e);
-        }
-    };
-
-    if shows.is_empty() {
-        eprintln!("⚠️  Warning: No shows found in file");
-        return Ok(());
-    }
-
-    println!("✅ Light show syntax is valid");
-    println!("   Found {} show(s):", shows.len());
-    for (name, show) in &shows {
-        println!("   - \"{}\" ({} cues)", name, show.cues.len());
-    }
-
-    // Get lighting config if provided
-    let (lighting_config, valid_groups_count, valid_fixtures_count) = if let Some(config_path) =
-        config_path
-    {
-        let config_file = Path::new(config_path);
-        if !config_file.exists() {
-            eprintln!("⚠️  Warning: Config file not found: {}", config_path);
-            (None, 0, 0)
-        } else {
-            match config::Player::deserialize(config_file) {
-                Ok(player_config) => {
-                    if let Some(dmx) = player_config.dmx() {
-                        if let Some(lighting) = dmx.lighting() {
-                            let groups_count = lighting.groups().len();
-                            let fixtures_count = lighting.fixtures().len();
-                            // Clone the lighting config to own it
-                            (Some(lighting.clone()), groups_count, fixtures_count)
-                        } else {
-                            eprintln!("⚠️  Warning: No lighting configuration found in DMX config");
-                            (None, 0, 0)
-                        }
-                    } else {
-                        eprintln!("⚠️  Warning: No DMX configuration found in config file");
-                        (None, 0, 0)
-                    }
-                }
-                Err(e) => {
-                    eprintln!("⚠️  Warning: Failed to parse config file: {}", e);
-                    (None, 0, 0)
-                }
-            }
-        }
-    } else {
-        (None, 0, 0)
-    };
-
-    // Use validation module to check groups
-    let validation_result = validate_groups(&shows, lighting_config.as_ref());
-
-    if validation_result.groups.is_empty() {
-        println!("⚠️  Warning: No fixture groups found in show");
-    } else {
-        println!("\n   Groups used in show:");
-        let mut sorted_groups: Vec<String> = validation_result.groups.iter().cloned().collect();
-        sorted_groups.sort();
-        for group in &sorted_groups {
-            println!("   - {}", group);
-        }
-    }
-
-    // Report validation results
-    if lighting_config.is_some() {
-        if !validation_result.is_valid() {
-            eprintln!("\n❌ Invalid groups/fixtures referenced in show:");
-            for group in &validation_result.invalid_groups {
-                eprintln!("   - {} (not found in config)", group);
-            }
-            return Err(format!(
-                "Show references {} invalid group(s)",
-                validation_result.invalid_groups.len()
-            )
-            .into());
-        } else {
-            println!("\n✅ All groups/fixtures are valid in config");
-            println!(
-                "   Validated against {} group(s) and {} fixture(s) in config",
-                valid_groups_count, valid_fixtures_count
-            );
-        }
-    }
-
-    Ok(())
-}
-
 #[tokio::main]
 async fn main() {
-    // Initialize tracing with a filter that sets default logging to off, with mtrack at info level
-    // This prevents noisy INFO messages from symphonia crates (which are suppressed by the default "off")
+    // Initialize tracing with a filter that sets default logging to off, with mtrack at info level.
+    // This prevents noisy INFO messages from symphonia crates (which are suppressed by the default "off").
     let filter = tracing_subscriber::EnvFilter::try_from_default_env()
         .unwrap_or_else(|_| tracing_subscriber::EnvFilter::new("off,mtrack=info"));
 
     tracing_subscriber::fmt().with_env_filter(filter).init();
 
-    if let Err(e) = run().await {
+    if let Err(e) = cli::run().await {
         eprintln!("Error: {}", e);
         std::process::exit(1);
     }
-}
-
-async fn run() -> Result<(), Box<dyn Error>> {
-    let cli = Cli::parse();
-
-    match cli.command {
-        Commands::Songs { path, init } => {
-            if init {
-                info!("Initializing songs");
-                songs::initialize_songs(Path::new(&path))?;
-            } else {
-                info!("Not initializing songs");
-            }
-
-            let songs = songs::get_all_songs(Path::new(&path))?;
-
-            if songs.is_empty() {
-                println!("No songs found in {}.", path.as_str());
-                return Ok(());
-            }
-
-            let mut all_tracks: HashSet<String> = HashSet::new();
-            println!("Songs (count: {}):", songs.len());
-            for song in songs.sorted_list() {
-                // Record all of the tracks found in the song repository.
-                for track in song.tracks().iter() {
-                    all_tracks.insert(track.name().to_string());
-                }
-
-                println!("- {}", song);
-            }
-
-            // Sort the tracks so that the output is consistent.
-            let mut tracks: Vec<String> = all_tracks.into_iter().collect();
-            tracks.sort();
-
-            println!("\nTracks (count: {}):", tracks.len());
-            for track in tracks.iter() {
-                println!("- {}", track)
-            }
-        }
-        Commands::Devices {} => {
-            print_device_list(audio::list_devices()?, "No devices found.");
-        }
-        Commands::MidiDevices {} => {
-            print_device_list(midi::list_devices()?, "No devices found.");
-        }
-        Commands::Playlist {
-            repository_path,
-            playlist_path,
-        } => {
-            let songs = songs::get_all_songs(Path::new(&repository_path))?;
-            let playlist = Playlist::new(
-                "playlist",
-                &config::Playlist::deserialize(Path::new(&playlist_path))?,
-                songs,
-            )?;
-
-            println!("{}", playlist);
-        }
-        Commands::Start {
-            player_path,
-            playlist_path,
-        } => {
-            apply_thread_priority();
-            let player_path = &Path::new(&player_path);
-            let player_config = config::Player::deserialize(player_path)?;
-            let mut playlist_path = playlist_path.as_ref().map(PathBuf::from).or(player_config.playlist()).ok_or(
-                "Must supply the playlist from the command line or inside the mtrack configuration",
-            )?;
-            if !playlist_path.is_absolute() {
-                playlist_path = if let Some(parent) = player_path.parent() {
-                    parent.join(playlist_path)
-                } else {
-                    return Err(format!(
-                        "Unable to determine playlist path (config base path has no parent): {}. \
-                         Please specify the playlist path using an absolute path.",
-                        playlist_path.display()
-                    )
-                    .into());
-                };
-            };
-            let songs = songs::get_all_songs(&player_config.songs(player_path))?;
-            let playlist = Playlist::new(
-                "playlist",
-                &config::Playlist::deserialize(playlist_path.as_path())?,
-                songs.clone(),
-            )?;
-
-            let player = Arc::new(player::Player::new(
-                songs,
-                playlist,
-                &player_config,
-                player_path.parent(),
-            )?);
-            crate::controller::Controller::new(player_config.controllers(), player)?
-                .join()
-                .await?;
-        }
-        Commands::Play { host_port, from } => {
-            let mut client = connect(host_port).await?;
-            if let Some(from_str) = from {
-                // Parse the time string
-                let start_time = parse_time_string(&from_str)?;
-                let start_duration = prost_types::Duration::try_from(start_time)
-                    .map_err(|e| format!("Failed to convert duration: {}", e))?;
-
-                let response = client
-                    .play_from(Request::new(PlayFromRequest {
-                        start_time: Some(start_duration),
-                    }))
-                    .await?;
-                println!("Playing the song from {}:", from_str);
-                print_song(response.into_inner().song)?;
-            } else {
-                let response = client.play(Request::new(PlayRequest {})).await?;
-                println!("Playing the song:");
-                print_song(response.into_inner().song)?;
-            }
-        }
-        Commands::Previous { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client.previous(Request::new(PreviousRequest {})).await?;
-            println!("Moved to previous song:");
-            print_song(response.into_inner().song)?;
-        }
-        Commands::Next { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client.next(Request::new(NextRequest {})).await?;
-            println!("Moved to next song:");
-            print_song(response.into_inner().song)?;
-        }
-        Commands::Stop { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client.stop(Request::new(StopRequest {})).await?;
-            println!("The song was stopped:");
-            print_song(response.into_inner().song)?;
-        }
-        Commands::SwitchToPlaylist {
-            host_port,
-            playlist_name,
-        } => {
-            let mut client = connect(host_port).await?;
-            let _ = client
-                .switch_to_playlist(Request::new(SwitchToPlaylistRequest {
-                    playlist_name: playlist_name.clone(),
-                }))
-                .await?;
-            println!("Switched to playlist {}", playlist_name);
-        }
-        Commands::Status { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client
-                .status(Request::new(StatusRequest {}))
-                .await?
-                .into_inner();
-            if let Some(song) = response.current_song {
-                println!("Current song: {}", song.name);
-                let song_duration = util::duration_minutes_seconds(Duration::try_from(
-                    song.duration.unwrap_or_default(),
-                )?);
-                let elapsed = Duration::try_from(response.elapsed.unwrap_or_default())
-                    .map(util::duration_minutes_seconds)?;
-                println!("Elapsed: {}/{}", elapsed, song_duration);
-            }
-            println!("Playing: {}", response.playing);
-            println!("Playlist name: {}", response.playlist_name)
-        }
-        Commands::ActiveEffects { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client
-                .get_active_effects(Request::new(GetActiveEffectsRequest {}))
-                .await?;
-            println!("{}", response.into_inner().active_effects);
-        }
-        Commands::Systemd {} => {
-            let current_executable_path = env::current_exe()?;
-            println!(
-                "{}",
-                SYSTEMD_SERVICE.replace(
-                    "{{ CURRENT_EXECUTABLE }}",
-                    current_executable_path
-                        .to_str()
-                        .expect("unable to convert current executable path to string")
-                )
-            )
-        }
-        Commands::VerifyLightShow { show_path, config } => {
-            verify_light_show(&show_path, config.as_deref())?;
-        }
-        Commands::Verify {
-            config,
-            check,
-            hostname,
-        } => {
-            let config_path = Path::new(&config);
-            let player_config = config::Player::deserialize(config_path)?;
-            let songs_path = player_config.songs(config_path);
-            let songs = songs::get_all_songs(&songs_path)?;
-
-            if songs.is_empty() {
-                println!("No songs found in {}.", songs_path.display());
-                return Ok(());
-            }
-
-            let run_all = check.is_none();
-            let checks: Vec<String> = check.unwrap_or_default();
-
-            let mut report = verify::VerificationReport::default();
-
-            // Track mapping checks.
-            if run_all || checks.iter().any(|c| c == "track-mappings") {
-                let all_profiles = player_config.all_profiles();
-
-                if all_profiles.len() > 1 {
-                    // Profile mode: verify each profile's track mappings.
-                    let profiles_to_check: Vec<&config::Profile> = match &hostname {
-                        Some(h) => {
-                            let filtered = player_config.profiles(h);
-                            if filtered.is_empty() {
-                                eprintln!("Warning: no profiles match hostname '{}'", h);
-                            }
-                            filtered
-                        }
-                        None => all_profiles.iter().collect(),
-                    };
-
-                    for (i, profile) in profiles_to_check.iter().enumerate() {
-                        let audio_config = match profile.audio_config() {
-                            Some(ac) => ac,
-                            None => {
-                                println!("Skipping profile {} (no audio configured)", i);
-                                continue;
-                            }
-                        };
-                        let label = match profile.hostname() {
-                            Some(h) => format!(
-                                "profile {} (hostname: {}, device: {})",
-                                i,
-                                h,
-                                audio_config.audio().device()
-                            ),
-                            None => format!(
-                                "profile {} (any host, device: {})",
-                                i,
-                                audio_config.audio().device()
-                            ),
-                        };
-                        println!("Checking track mappings for {}...", label);
-                        let track_report =
-                            verify::check_all_track_mappings(&songs, audio_config.track_mappings());
-                        report.merge(track_report);
-                    }
-                } else {
-                    // Single profile (legacy or single profile): verify as before.
-                    let track_report =
-                        verify::check_all_track_mappings(&songs, player_config.track_mappings());
-                    report.merge(track_report);
-                }
-            }
-
-            verify::print_report(&report, &songs);
-
-            if report.has_errors() {
-                std::process::exit(1);
-            }
-        }
-        Commands::Cues { host_port } => {
-            let mut client = connect(host_port).await?;
-            let response = client
-                .get_cues(Request::new(GetCuesRequest {}))
-                .await?
-                .into_inner();
-
-            if response.cues.is_empty() {
-                println!("No cues found in the current song.");
-            } else {
-                println!("Cues in current song ({} total):", response.cues.len());
-                for cue in response.cues {
-                    let time = cue
-                        .time
-                        .and_then(|d| Duration::try_from(d).ok())
-                        .map(util::duration_minutes_seconds)
-                        .unwrap_or_else(|| "unknown".to_string());
-                    println!("  {}: {} (index {})", cue.index, time, cue.index);
-                }
-            }
-        }
-    }
-
-    Ok(())
-}
-
-fn print_song(song: Option<Song>) -> Result<(), Box<dyn Error>> {
-    if let Some(song) = song {
-        println!("Name: {}", song.name);
-        println!(
-            "Duration: {}",
-            util::duration_minutes_seconds(Duration::try_from(song.duration.unwrap_or_default())?)
-        );
-        println!("Tracks:");
-        for track in song.tracks {
-            println!("  - {}", track);
-        }
-    }
-
-    Ok(())
-}
-
-async fn connect(
-    host_port: Option<String>,
-) -> Result<PlayerServiceClient<Channel>, Box<dyn Error>> {
-    // Use 127.0.0.1 for local client; 0.0.0.0 is a bind address, not valid for connect.
-    let addr = host_port.unwrap_or_else(|| format!("127.0.0.1:{}", config::DEFAULT_GRPC_PORT));
-    Ok(PlayerServiceClient::connect(format!("http://{}", addr)).await?)
 }
